+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 * thickness to `Boxes.draw_on`
+* add `round` to Boxes, Points, and Coords
+* add `contains` to Boxes
 
 
 ## Version 0.5.7 - Released 2020-January-23